+++ src/components/Pipeline.vue
@@ -34,6 +34,7 @@
         Delete [{{ selectedSteps.length }}] selected
       </div>
     </div>
+    <PreviewSourceSubset />
     <div class="query-pipeline__tips-container" v-if="hasSupportedSteps">
       <div class="query-pipeline__tips">
         Interact with the widgets and table on the right to add steps
@@ -62,6 +63,7 @@ import { VQBModule } from '@/store';
 import { MutationCallbacks } from '@/store/mutations';
 
 import DeleteConfirmationModal from './DeleteConfirmationModal.vue';
+import PreviewSourceSubset from './PreviewSourceSubset.vue';
 import Step from './Step.vue';
 
 @Component({
@@ -71,6 +73,7 @@ import Step from './Step.vue';
     Draggable,
     Step,
     FAIcon,
+    PreviewSourceSubset,
   },
 })
 export default class PipelineComponent extends Vue {